+++ src/AtomicAndPhysicalConstants.jl
@@ -22,9 +22,8 @@ include("CODATA2018.jl")
 include("CODATA2022.jl")
 include("constants.jl")
 include("helpers.jl")
-include("atomic_species.jl")
-include("subatomic_species.jl")
-include("species.jl")
+include("species_data.jl")
+include("constructors.jl")
 # include("gyro_vals.jl")
 include("getter_functions.jl")
