--- src/subatomic_species.jl
@@ -1,7 +1,0 @@
-
-# AtomicAndPhysicalConstants/src/subatomic_species.jl
-
-
-
-
-
